test(server): split readonly auth and header policy tests
